+++ release-notes.md
@@ -12,6 +12,7 @@ Version 1.4.1 (unreleased)
 * Added Vec3d.distance(x,y,z) and Vec3d.distanceSq(x,y,z)
 * Added Vec3i.length() and Vec3i.lengthSq()
 * Added a Vec3d.ZERO 'constant'
+* Added Rayd, a double-precision Ray class.
 
 
 Version 1.4.0 (latest)